Upgrade all execute steps to use new step -> items structure
