+++ packages/client/src/types/index.ts
@@ -2,26 +2,6 @@ import { paths } from './api'
 export * from './api'
 
 export type Execute = {
-  quote?: number
-  steps?:
-    | {
-        action: string
-        description: string
-        status: 'complete' | 'incomplete'
-        message?: string
-        error?: string
-        kind: 'transaction' | 'signature' | 'request' | 'confirmation'
-        data?: any
-        txHash?: string
-      }[]
-  query?: { [key: string]: any }
-  statusCode?: number
-  error?: string
-  message?: string
-  txHash?: string
-}
-
-export type BatchExecute = {
   error?: string
   path:
     | paths['/execute/buy/v3']['get']['responses']['200']['schema']['path']
