+++ packages/client/src/utils/index.ts
@@ -1,5 +1,4 @@
 export { executeSteps } from './executeSteps'
-export { batchExecuteSteps } from './batchExecuteSteps'
 export { setParams } from './params'
 export { pollUntilOk, pollUntilHasData } from './pollApi'
 export { isOpenSeaBanned } from './isOpenSeaBanned'
